==> explore_figure/Fig7_Tcell/05_Heatmap_Idents_submain.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(tidyverse)
library(here)
library(SingleCellExperiment)
library(cowplot)
library(knitr)
library(sctransform)
library(Matrix)
library(future)
library(future.apply)
```

```{r}
tcell <- readRDS(here('explore/RDS/Tcell2.RDS'))
```

```{r}
DefaultAssay(tcell) <- "integrated"
Idents(tcell) <- "new_submain2"
```

```{r}
mast_markers <- FindAllMarkers(tcell, only.pos = T, min.pct = 0.25, logfc.threshold = 0.25, test.use = "MAST")
```


```{r}
top20 <- mast_markers %>% 
  group_by(cluster) %>% 
  top_n(n = 20, wt = avg_log2FC)
top20
```

```{r}
DefaultAssay(tcell) <- "integrated"

DoHeatmap(tcell, features = top20$gene) + NoLegend() 
```
