==> explore_figure/Fig7_Tcell/07_Subpopulation_RDS.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(tidyverse)
library(here)
library(SingleCellExperiment)
library(cowplot)
library(knitr)
library(sctransform)
library(Matrix)
library(future)
library(future.apply)
```

```{r}
tcell <- readRDS(here('explore/RDS/Tcell2.RDS'))
```

```{r}
DimPlot(tcell, group.by = "seurat_clusters", label = T)
```


```{r}
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(0, 7, 5), "Tcell_SC1", "NA")
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(11), "Tcell_SC2", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(12), "Tcell_SC3", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(1, 15), "Tcell_SC4", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(2), "Tcell_SC5", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(12), "Tcell_SC6", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(6), "Tcell_SC7", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(16, 14), "Tcell_SC8", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(4, 18, 19), "Tcell_SC9", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(13), "Tcell_SC10", tcell@meta.data$new_submain3)

tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(10, 8), "NKcell_SC1", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(3, 20), "NKcell_SC2", tcell@meta.data$new_submain3)
tcell@meta.data$new_submain3 <- ifelse(tcell@meta.data$seurat_clusters %in% c(9), "NKcell_SC3", tcell@meta.data$new_submain3)



DimPlot(tcell, group.by = "new_submain3", label = T, repel = T)
```

```{r}
saveRDS(tcell, file = "Tcell3.RDS")
```
